+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## v1.4.0 - 2021-12-12
 
-- Handle RSS as a format in addition to Bookmarks Json format
+- Handle RSS and Atom as formats in addition to Bookmarks Json format
 
 ## v1.3.2 - 2020-05-08